6/3/24
I made it so the red bullet shoots and has a cooldown(fixed bugs with it), i made it so the bullet one taps the boss and i also added that if the boss ship collides with the base, the game ends. I also added a manuel type thing where it shows controls before you start the game.